+++ CHANGES.rst
@@ -8,6 +8,7 @@ Development
 - Explicitly ``deinit()`` LTE modem on each startup
 - Fix ``settings.example-bob.py``. Thanks, @MKO1640 and @ClemensGruber!
 - Improve formatting of BEEP telemetry field mapping for BOB
+- Disable telemetry adapter offline state for now
 
 
 2019-06-22 0.5.1